+++ app/utils/transform-query-params.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-transition-helper/utils/transform-query-params';