+++ tests/dummy/app/routes/double.js
@@ -1,0 +1,7 @@
+import Route from '@ember/routing/route';
+
+export default class TestRoute extends Route {
+  model(params) {
+    return params;
+  }
+}